+++ main.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+import os
 import sys
 sys.path.append('/usr/local/google_appengine/')
 
@@ -8,13 +9,15 @@ import mail.auth
 import mail.check
 import mail.list
 
+DEBUG = os.environ.get('APP_DEBUG') == '1'
+
 app = webapp2.WSGIApplication([
     ('/auth', mail.auth.Handler),
     ('/mail/check', mail.check.Handler),
     ('/mail/list', mail.list.Handler),
     ('/mail/me', mail.list.Single)
-], debug=True)
+], debug=DEBUG)
 
 
 if __name__ == '__main__':
-    unittest.main()
+    print('Run this application with the App Engine development server.')